requirements.txt
requests
beautifulsoup4
python-telegram-bot
scikit-learn
pandas
numpy
nltk
